fix(app): fixed style flash in post-form page
