+++ src/app/postForm/layout.tsx
@@ -1,0 +1,6 @@
+import SingleComponentLayout from "@/modules/layouts/singleComponentLayout"
+import { ReactNode } from "react"
+
+export default function layout({ children }: { children: ReactNode }) {
+  return <SingleComponentLayout>{children}</SingleComponentLayout>
+}
